requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
yfinance==0.2.51
matplotlib==3.9.4
